pyproject.toml
[project]
name = "invest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.17.44",
    "jupyter>=1.1.1",
    "mplfinance>=0.12.10b0",
    "pandas>=2.3.2",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.2",
    "tqdm>=4.67.1",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
